Fix numpy type definitions

+++ oxasl_ve/api.py
@@ -64,7 +64,7 @@ def veslocs_to_enc(veslocs, nvols=8):
                 [tag_deg, 3, vA, vB],
             ]
 
-        return np.array(two, dtype=np.float)
+        return np.array(two, dtype=np.float32)
     except: 
         traceback.print_exc()
         raise RuntimeError("Error generating encoding matrix from vessel locations. Check your vessel locations file.")
@@ -110,7 +110,7 @@ def two_to_mac(two):
             inc += 1
 
     mac = [cx, cy, th, D]
-    return np.array(mac, dtype=np.float), imlist
+    return np.array(mac, dtype=np.float32), imlist
 
 def mac_to_two(mac):
     """ 
@@ -188,7 +188,7 @@ def generate_mask(data, imlist, frac=0.5):
     tag_idx, ctl_idx = imlist.index(-1), imlist.index(0)
     diffdata = np.abs(data[..., tag_idx] - data[..., ctl_idx])
     thresh = np.percentile(diffdata, 99) * (1-frac)
-    return (diffdata > thresh).astype(np.int)
+    return (diffdata > thresh).astype(np.int32)
 
 def _autogen_mask(wsp):
     maskdata = generate_mask(wsp.asldata.data, wsp.imlist, wsp.ifnone("infer_mask_frac", 0.5))
@@ -266,7 +266,7 @@ def _decode(wsp):
     wsp.veasl.infer_loc = wsp.ifnone("infer_loc", "rigid")
     if wsp.init_loc and wsp.asldata.ntis > 1:
         wsp.log.write("\n - Doing initial fit for vessel locations using mean data\n")
-        asldata_mean = np.zeros(list(wsp.veasl.asldata_mar.data.shape[:3]) + [wsp.asldata.ntc], dtype=np.float)
+        asldata_mean = np.zeros(list(wsp.veasl.asldata_mar.data.shape[:3]) + [wsp.asldata.ntc], dtype=np.float32)
         for idx in range(wsp.asldata.ntis):
             asldata_mean += wsp.veasl.asldata_mar[..., idx*wsp.asldata.ntc:(idx+1)*wsp.asldata.ntc]
         asldata_mean /= wsp.asldata.ntis
@@ -310,7 +310,7 @@ def _model_vessels(wsp, num_vessels):
     for vessel in range(num_vessels):
         wsp.log.write("\n - Processing vessel %i\n" % (vessel+1))
         wsp_ves = wsp.veasl.sub("vessel%i" % (vessel+1))
-        vessel_data = np.zeros(list(wsp.asldata.shape[:3]) + [wsp.asldata.ntis,], dtype=np.float)
+        vessel_data = np.zeros(list(wsp.asldata.shape[:3]) + [wsp.asldata.ntis,], dtype=np.float32)
         for ti_idx in range(wsp.asldata.ntis):
             flow = getattr(wsp.veasl, "pld%i" % (ti_idx+1)).flow
             vessel_data[..., ti_idx] = flow.data[..., vessel] * 2
@@ -336,7 +336,7 @@ def _combine_vessels_sum(wsp, num_vessels, output):
         vessel_img = getattr(vessel_wsp.native, output, None)
         if vessel_img is not None:
             if all_vessel_img is None:
-                all_vessel_img = np.zeros(vessel_img.shape, dtype=np.float)
+                all_vessel_img = np.zeros(vessel_img.shape, dtype=np.float32)
             all_vessel_img += vessel_img.data
             have_output = True
     if have_output:
@@ -349,8 +349,8 @@ def _combine_vessels_weighted(wsp, num_vessels, output, method="weightedperf"):
     
     Used for arrival time and variance/std dev
     """
-    vessel_perf = np.zeros(list(wsp.asldata.shape[:3]) + [num_vessels,], dtype=np.float)
-    vessel_output = np.zeros(list(wsp.asldata.shape[:3]) + [num_vessels,], dtype=np.float)
+    vessel_perf = np.zeros(list(wsp.asldata.shape[:3]) + [num_vessels,], dtype=np.float32)
+    vessel_output = np.zeros(list(wsp.asldata.shape[:3]) + [num_vessels,], dtype=np.float32)
     have_output = False
     for vessel in range(num_vessels):
         vessel_wsp = getattr(wsp.output, "vessel%i" % (vessel+1))
